Refactor EntityFactory and JSON converters
Updated `EntityFactory` to support additional entity types (`Polyline3D`, `PolyfaceMesh`, `PolygonMesh`) and refactored randomization methods. Simplified `RandomizePolyline<T>` with generic constraints.

Introduced `writeEnumerable` in `CommonCadConverter` for enumerable serialization, supporting both `System.Text.Json` and `Newtonsoft.Json`. Moved `CommonPolylineConverter<T>` to a new file with updated logic.

Removed `CSUtilities.Extensions` references and cleaned up conditional compilation. Performed general code cleanup and reorganization for better maintainability.
